POA refactoring
Move the PoaGraphImpl into a separate file, and move the graph
traversals into another file.

[Tracking bug 24842]

[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/lib/cpp/ConsensusCore/": change = 146251]
